Synchronize char creation wizard screens
Addresses a bug described here: https://forums.dfworkshop.net/viewtopic.php?p=66200

The skill bonus screen de-synchronizes with the summary screen when the user goes back and forth in the character creation wizard. My fix was to ensure either window is synced to the character document whenever they are pushed. I also made it so that changes made on the summary page are propagated backward in the wizard as far as makes sense. Skills bonuses will still be wiped if the user creates a new biography since that alters them.