Disable collider on MagicCandleBehaviour billboard
Any mod registering a custom activation to candle sprite will result in a collider on magic candle billboard.
This can block spells and be a nuisance to player.
MagicCandleBehaviour will now disable collider on children at start. This only happens when effect is instantiated, such as first cast or loading a game.
There should only be a single collider on child billboard in this scenario.