Allow repeated IME composition input
The intent of removed code was to prevent creation of excess string garbage, but it would also prevent entering same single-character composition in a row.
It's not possible to know what IME provider is actually doing as this happens at OS level, so we need to assume live update for inline composition preview.
This can possibly be refined later.