Check IME disabled anytime game unpaused
IME composition will break game input if enabled during play.
This shouldn't happen in the course of normal play, but might happen if UI crashes or a window failed to restore IME state on close.
This is a kind of soft recovery to disable IME composition anytime game is unpaused and control returns to player.