talk_print console command
Helper to output talk text to console by ID. Tokens are routed through TalkManager MCP. When there are multiple sub-records for this ID, one will be selected at random.
Using this to debug a localization issue. Adding to master as it might come in handy for someone else in the future.