Slightly increase Mage Light range and intensity
It looks like this example spell is sticking around. So try to make it more equal to magic candle in usefulness.
As candle appears in front of player and mage light is centred on player, make mage light radius a bit larger so it casts ahead of player an equal distance to magic candle.
As mage light has a soft coloured glow, slightly increase intensity.