Fix regionIndex not passed
regionIndex is required to correctly lookup localized region name for macro expansion. This was not passed when collecting site buildings for quest targets. Notably this would send player to a location in Alik'r Desert (not initialised regionIndex=0) instead of Wayrest as expected.
regionIndex is now passed for all new SiteDetails() in Place.