Upgrade DFU to 2022.2.21f1 LTS
I don't expect this to be merged, it's just to get the ball rolling on updating to a new Unity LTS version in future.

Process:
* Upgraded project to 2022.2.21f1 (ignore warnings).
* Removed Jobs package (old experimental package no longer required).
* Removed some /Logs folder spam.

Everything looks to play OK after upgrade with no significant problems or errors found at this time in base game. Obviously not testing mods yet, and many of them will probably need to be rebuilt for new engine version.

With the usual preview lead and communication for modders to update their builds, the project can be upgraded to a new engine version if required.