Move the live change warning outside of the settings panel
The warning for changing live settings being inside the panel messes up the careful formatting of the mod settings. I know that for my mods I spend a lot of time getting them looking good.

I also think the message is more obvious not being lost in all of the settings text.