Fix spell absorption capping
When absorbing magicka from own spells, DFU explicitly ensures that
magicka recovered is not larger than the magicka that was used.
For this, it needs to track how much magicka was spent for the last
spell: lastReadySpellCastingCost

A special case was not correctly handled though: when the last spell
came from a magic item, lastReadySpellCastingCost should be set 0 so
that next absorption of own spells is not capped.
That worked by luck when the game was just started because
lastReadySpellCastingCost was still 0, but broke after first spell
casting.

Forums: https://forums.dfworkshop.net/viewtopic.php?t=6755