Allow to close console with Escape key
Saw a streamer struggle twice because he opened the console
unintentionnally and had no idea how to close it, so closed and
restarted the game instead.

Unless there's any inconvenience, I think it would be best to enable by
default an already existing feature of the console: allow to close it
with the Escape key in addition to the console toggle key.

(some additionnal fixes required because of bit rot)