Fix and improve icons layouts
One small issue found after fixing UI positioning, is that in 4:3
resolution the first debuff icon and the action mode icon overlap.

Going back to classic Daggerfall, the first debuff icon is actually to
the right of the first buff icon (in fact it's aligned with the 2nd buff
icon).

Matching classic actually fixes that small issue.

What I was at it, improved the way icons adjust to the HUD by specifying
whether they're anchored to the top, bottom, left or right of the
screen.