Add More Substantial Override Paths For Physical Combat, 9/11/2024
Add More Substantial Override Paths For Physical Combat. This idea of these changes is to allow for much more of the vanilla behavior to be circumvented if a modder chooses to do their own thing, in this instance hopefully with as little impact on other parts of the code and execution as possible, 9/11/2024.