Big rewrite of "Who Gets the Totem"
The problem this solved was the oddity of trying to surrender the totem to Gothryd's mercenaries. This required reworkings of several quest components. Tested about 20 times through, looks solid now. Also added a few QOL features such as an updated log entry when the totem is given, and a failure message when the year timer has passed. Also forces spawn of Eadwyre's mercenaries immediately instead of taking 25 seconds.