DaggerfallHUD: Use 320x240 NativePanel
With 320x200 activeSpells panel, icons can fall below the panel and bad
things occur (say missing tooltips).
Given LocalScale, activeSpells, like the NativePanel, need to be 320x240
to cover the whole screen in narrow resolutions, but it's unclear for me
if that's intended, or if they're supposed to be 320x200 but stretched,
or even something else.